requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
alembic
pydantic
python-dotenv
celery[redis]
redis
requests
sentence-transformers
numpy
scipy
openai
python-jose
passlib
faiss-cpu
python-multipart